Add timeline and graph views with tests
