+++ frontend/package.json
@@ -7,9 +7,11 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview --port 5173 --strictPort",
-    "postinstall": "npx playwright install --with-deps"
+    "postinstall": "npx playwright install --with-deps",
+    "test": "vitest run"
   },
   "dependencies": {
+    "d3": "^7.9.0",
     "maplibre-gl": "^4.5.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
@@ -17,9 +19,12 @@
   "devDependencies": {
     "@playwright/test": "^1.55.0",
     "@vitejs/plugin-react": "^4.3.1",
+    "@testing-library/react": "^14.2.2",
+    "jsdom": "^24.1.0",
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
     "typescript": "^5.5.4",
-    "vite": "^5.4.2"
+    "vite": "^5.4.2",
+    "vitest": "^2.1.1"
   }
 }
